improve updating of the year_of_activity
Instead of keeping a flag of is_active_today (which no longer matters if the
cron doesn't get run every day, or is out of sync), this records the
timestamp of each time the year_of_activity is updated.

Therefore, if the cron failed to run (or any other such problem), we would
still know how many days since the last activity investigation, and be able to
update just those days.