moved to a DummySync for all models
We want to implement Syncable, for the destroy() and toJSON() methods,
and that requires we add a Sync to use them (else destroy() will throw an
error trying to find the sync to use).

We also don't want to currently use the ServerSync, since the new API isn't
in place. So Dummy it is!