Add 'permissions' to allowed keys.
SDK 1.14 will add a 'permissions' key to package.json, initially used only for allowing addons to opt-in to running in private browsing windows, with other potential uses in the future. This commit adds support for the key so developers don't get errors when trying to add the key to their addons.